add dropdown to search bar componenet
